pyproject.toml
[tool.poetry]
name = "midmarketapp"
version = "0.1.0"
description = "API for getting the Mid Market rate for any currency"
authors = ["Delmwin Baeka <dbaekajnr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.89.1" }
celery = { extras = ["redis"], version = "^5.2.7" }
passlib = { extras = ["bcrypt"], version = "^1.7.2" }
pydantic = "^1.4"
gunicorn = "^20.0.4"
sqlalchemy = "^1.3.16"
python-jose = { extras = ["cryptography"], version = "^3.1.0" }
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
